adds RepoList component
